pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rclone-bisync-manager"
version = "0.1.0"
description = "A daemon-based solution for automated, bidirectional synchronization of files using RClone"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "croniter==3.0.3",
    "pydantic==2.8.2",
    "python_daemon==3.0.1",
    "PyYAML==6.0.2",
    "psutil==5.9.5",
]

[project.scripts]
rclone-bisync-manager = "rclone_bisync_manager.main:main"
rclone-bisync-manager-tray = "rclone_bisync_manager_tray.system_tray:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["rclone_bisync_manager*", "rclone_bisync_manager_tray*"]

[project.optional-dependencies]
tray = [
    "Pillow==10.4.0",
    "pystray==0.19.5",
    "PyGObject==3.46.0",
    "CairoSVG==2.7.1",
    "PyYAML==6.0.2",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/rclone-bisync-manager"
"Bug Tracker" = "https://github.com/yourusername/rclone-bisync-manager/issues"